+++ changelog.rst
@@ -5,6 +5,7 @@ Release 0.3.13 (not released yet)
 ---------------------------------
 
 - Fix: restart policy cannot be applied to temporary containers (`#95`_)
+- Enhancement: ``DockerTasks.deploy()``: added option `prepare=True` which can be used to skip 'prepare' step during deploy
 
 .. _#95: https://github.com/renskiy/fabricio/issues/95